Rename misspelled vendormodule
