+++ vendormodule.js
